Modularize chmod helpers
